Update patch set 2
Patch Set 2: Code-Review+1

@Sylvain,as Mukesh explained in his email to you...


This is in reference to your review https://gerrit.onap.org/r/c/oom/+/114367/4/kubernetes/so/components/so-vnfm-adapter/resources/config/overrides/override.yaml
 
The Ericsson team has discussed internally about the above hardcoded password. The reason that we put keystore and truststore passwords in the SOL003 Adapter is that the SOL003 Adapter has used its own keystore for communications with bpmn-infra and etsi-nfvo.
 
If we are using the common way of certificates as you are changing, it will break existing communications between bpmn-infra/etsi-nfvo and SOL003 Adapter.
 
As we reported it to you, due to above changes in OOM, we are not able to successfully test the communications around the SOL003 Adapter.
 
Since we don’t have much time to complete our tests for ETSI-Alignment (it used to work but after OOM changes, no), at this time, out team would like to go with the HTTP path instead of HTTPS, as part of this review for Guilin. Another rationale for the HTTP path would be the SO NFVO, SOL003 Adapter and BPMN-Infra are part of SO. In Guilin ,we can treat their communications as SO internal, where HTTPS support is not mandatory. For Guilin, this HTTP path would simplify current issues around the SOL003 Adapter.
 
Security is important. In Honolulu, we will revisit these security work items, as part of https://jira.onap.org/browse/SO-3350 epic. Thanks.

Could you please let us know your view and advise. 

BR,
Byung

Patch-set: 2
Reviewer: Gerrit User 2085 <2085@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1