Update patch set 18
Patch Set 18: Code-Review-1

Removing the values.yaml for so-appc-adapter is insufficient.
A complete Helm chart shall be removed as well as requirements.yaml in so need to be modified, and the onap/values.yaml.

You can have a look at this one: 
https://gerrit.onap.org/r/c/oom/+/114431 

As a removal of a Helm chart is a bigger thing, I would propose to do this in a dedicated MR, rather than this MR. We shall simply not cover a change for that affected Helm Chart...

Patch-set: 18
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1