Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Sylvain: Could you pls merge it?
> 
> onap-vnfsdk-5f758494b8-dsptq 	1/1 	Running 	None 	0
> onap-vnfsdk-init-postgres-k2shr 	0/1 	Succeeded 	None 	0
> onap-vnfsdk-postgres-primary-5f64f9596c-qfzch 	1/1 	Running 	None 	0
> onap-vnfsdk-postgres-replica-685566d467-xpn88 	1/1 	Running 	None 	0

we need to wait for verify job.
And we (Krzsysztof and myself) are trying to regularly check the status of the merge request.
So you don't need to remind it to us, unless it's more than one day of wait

Patch-set: 1
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>