Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Verified manually on a recent master branch:
> > [ERROR] Only release images are allowed in helm charts.
> > [ERROR] Images not found in release repo:
> > onap/org.onap.dcaegen2.deployments.pnda-bootstrap-container:6.0.0
> > onap/org.onap.dcaegen2.deployments.pnda-mirror-container:6.0.0
> > onap/so/so-etsi-nfvo-ns-lcm:1.7.4
> > Makefile:101: recipe for target 'check-for-staging-images' failed
> > make: *** [check-for-staging-images] Error 1
> > 
> > results are aligned with expectations
> 
> I propose to start to work on adding to jjb but to merge (the jjb change) only when we have solved the initial problem (so no more "staging" images first)

My idea was to create a second jenkins job so that it doesn't prevent full build even through there are some staging images and for now just merge patches even if this job is failed.

This way I would not need to run this manually on dozen of reviews every day...
What do you think?

Patch-set: 7