Update patch set 3
Patch Set 3: Code-Review+2

@Sandeep,

my change was very minor (add [PORTAL] in front of first sentence, remove portal in the text of this sentence) and your commit message was 95% fine :)

please note that first line shouldn't be longer than 50 characters as it's a best practice (but here again you were good)!

we mandate this format of commit message:
[NAME_OF_COMPONENT|DOC|COMMON|GENERIC] Meaningful title (from OOM side)

at least one sentence explaining the change done in this patch, cause and consequences and possibly more of course

Issue-ID: AS_WE_ARE_FORCED_BUT_MEANINGLESS
Change-ID: xxx
Sign-off: xxx

Commit message will be the last stuff that will stay with our code so it must clearly explain the changes, the "why" and the consequences. If it change OOM behavior in any way, documentation must be also updated.

Starting Frankfurt branching, merge requests which are not following this pattern will not be merged.

Please read the following pages and follow the guidelines for writing commit message contained therein.

    http://bit.ly/goodcommitmessages
    http://who-t.blogspot.com/2009/12/on-commit-messages.html
    http://dep.debian.net/deps/dep3/

Patch-set: 3
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2