[Tree-wide] Update helm repo after build
Helm 3 no longer supports creation of a local repo as helm 2 did. Now
we need to use external component named chartmuseum to provide us repo
for helm charts.

When we start chartmuseum for the first time repo is empty and during
our build process we successively built helm charts to it. As a last
chart we always build our umbrella chart called onap and also push it
to repo. Unfortunately our local helm cache is unaware that this new
chart is now available untill we push sth else or execute helm repo
update which leads to really cryptic deployment failures.

To mitigate this issue let's add a new target helm-repo-update which
will be executed right after onap and just make sure that our local
cache is up to date.

WARNING:
If you are using helm v3 and you build a single component for the
first time on this machine or version of chart for the component that
you are building has just changed recently it's not enough to run:

$ make <component>

You need to also run:

$ make helm-repo-update

after successful build of your component.

Issue-ID: OOM-2562
Change-Id: I77020a3fb5666106b86c6a76477a57d9dd5af047
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>