Update patch set 1
Patch Set 1:

The Jira used in the Issue-ID is for dual stack support, which doesn't seem right.  Is this upgrade truly needed to support dual stack, or is this to mediate a security vulnerability?

In either case, this seems quite late for Guilin, unless it is needed to resolve a blocking issue.  I'm absolutely fine with this in Honolulu, though.

Patch-set: 1
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>