Update patch set 26
Patch Set 26:

> Patch Set 26: Code-Review-1
> 
> basic_cnf failed sorry:
> [SO] YamlTemplateVfModuleAlaCarteInstantiateStep: Instantiate VF module described in YAML using SO a'la carte method.

according to the logs https://logs.onap.org/onap-integration/gating//113937-26/infrastructure-healthcheck/k8s-teardown/kubernetes-status/pod-onap-multicloud-k8s-c464bdcf7-48bft-multicloud-k8s.log
the issue is due to the fact that the resources we want to create already exist

in theory the test should have cleaned it, but we already remarked that it is not always the case

so we planned to add a system cleanup to be sure (in addition of a test cleanup...). But it is not available yet (in progress)

the issue is then maybe due to the test conditions. We do not have in daily because...we reinstalled the full k8s, in gating to save time we just reinstall the ONAP cluster

I would be in favor of a +1

as we have the night ahead, we can give a new try, and I can manually delete the resources on the target lab..to be sure

oom_redeploy

Patch-set: 26
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>