Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > yes as far as I understand the script https://github.com/onap/integration-xtesting/blob/master/security/scripts/check_security_root.sh
> > > 
> > > we should not retrieve retrieved something like 
> > > kubectl exec onap-msb-consul-5d44fffcdb-lfgr4 --container msb-consul -n onap id
> > > uid=0(root) gid=0(root) groups=0(root),1(bin),2(daemon),3(sys),4(adm),6(disk),10(wheel),11(floppy),20(dialout),26(tape),27(video)
> > > what I got today
> > > we should have uid=1(consul)
> > > 
> > > oom_redeploy
> > 
> > We'll Then maybe the test script is imperfect. So the entry point starts as a root but then it drops root privileges and executes main application as a root. That's how the upstream image was constructed...
> 
> 
> 
> The script is simple, we connect to each pod, and run id, if the id =0 / root we rise the flag.
> ROOT is prohibited inside pod.

So in my opinion what is prohibited are processes running as root not that the entrypoint starts as root which this actually checks.

So the way consul upstream container works is that it starts as a root does some prep and then drops root capabilities before starting the actual consul service (Yes, this should have been done in Dockerfile but don't blame me for that it's the way consul upstream does it).

So after this start there is no process running as a root within the container so even if the consul service is hacked it shouldn't be able to gain root privileges (unless they have LPE exploit). WHen you exec into this container you start a new process with the same privilege level as entrypoint started so effectively root and that's why this test fails even though we don't have any root process in this container unless you spawn one.

Is it an issue? It all depends on the threat model. If we want to protect our worker node from running unknown code as a root they yes. But if we assume that containers that we are running are certified and safe and we just want to protect against malicious attacker who hacked our consul service then it doesn't

Patch-set: 1