Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Krzysztof, your previous commit introduce some kind of DoS attack into ONAP. Consule generates a lot of defunct processes and kills whole worker because OS cannot open any new processes. After a while consules starts on other worker and situation starts again and next worker were killed. Please check what is corrected in your commit and merge it asap.

Daniel,
I'm not negating that I've introduced a bug. All what I say is that:
a) I've not observed this thus I'd be more than happy to get more details. ie list of zombie processes (when I run it locally I've none). For an issue like this I'd expect a very precise and descriptive error message. If I could get more details on where are the zombie processes coming from maybe we could find a better solution than reintroducing root process

b) It's not the proper way of fixing this issue as it reintroduces a bug that we've already fixed.

We can either workout a proper fix for this or instead of silently reintroducing the issue just send a revert for my commit which would be more proper than this patch.

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>