Update patch set 1
Patch Set 1:

Now I see that my fix reverts to execute from root but still as long we are staying in this situation it is blocker for our work. Can we merge it as temporary solution to fix master and after that you introduce the right solution for non-root user?

Patch-set: 1