Update patch set 1
Patch Set 1:

Breaking the cite line to avoid triggering redeploy once again.

> yes it is an issue, you leave the possibility to gain privilege if the root account still running.

There is no such thing as "root account still running" accounts are not running but only processes does that. That's why I wrote that if we consider consul container to be secure then there is no security issue as there is just a single process inside the container and it runs as a consul user so even if you hack it you would need LPE exploit to gain access but it looks pretty much the same if no matter if your entrypoint starts as root or not.

Think how docker works. There is a docker daemon that runs as a root. When you want to start the container this daemon executes fork(), calls unshare() to move to a separate namespace, sets up cgroups, drops priviledges and then either switch user to the one that you requested or keeps it running as root.

So the only difference in situation where docker daemon runs entrypoint as non-root and the entrypoint itself switch user are the few lines of code from the entrypoint that are executed with root permissions. So *as long as the entrypoint code is trusted* there should be no issues with that

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>