Update patch set 16
Patch Set 16:

OK we fixed it. The problem was bad permission of mounted(binded) directory on container. It had user and group of "1001", but default user on docker(which executed all the commands, like create file or generate certificates) was "onap". Now we set default user to root. And it's working.

Patch-set: 16
Work-in-progress: false