Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-2
> 
> -2 until the patch is merged into master
> We strongly encourage to cherry-pick patches only *after* they are merged into master.
> This avoids confusion and saves gating resources

I did this on Friday after correction of all your -1 in order to have gating results for both commit and cherry pick. Also I checked installation of SDC with this change on my Guilin lab. 
I assumed that gating resources are less utilized on the weekend and what is more, more stable and reliable.

Patch-set: 2