Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> > Patch Set 1:
> > 
> > @Morgan
> > Any idea on the nature of basic_network failure?
> 
> here it is a dahboard issue
> because both
> basic_network and basic_cnf are PASS if you click on the link..
> 
> so it is only a dashboard issue, maybe results were not pushed in the DB
> the dashboard status is built dynamically based on what can be retrieved from the DB
> so artifacts can be good but if DB status is not good => dashboard will be like that
> usually it is sync..
> 2 possible reasons
> - results were not pushed
> - results were pished with the same CI id, dashboard take the first one (faulty)
> need to dig a little bit..

very strange
dashboard red because it is reported as failed in the DB
test report is green https://logs.onap.org/onap-integration/gating//115243-1/smoke-usecases/basic_network/reporting.html
but the link inside the reporting lead to an error of SDC distribution...(from the 23/11)
https://logs.onap.org/onap-integration/gating//115243-1/smoke-usecases/basic_network/pythonsdk.debug.log

regarding the job
https://gitlab.com/Orange-OpenSource/lfn/onap/xtesting-onap/-/jobs/871106709
it says something else 
Unknown test case or tier 'basic_network' seen in the logs..

As far as I understand
the retry today was done with an "old xtesting" docker (LF restarted some servers last night? no idea why an old version was considered, I need to check on recent gating if it is still the case)
this old docker was not including basic_network and basic_cnf (not executed) => error
this explains that the test in DB is FAIL and the dashboard is red

BUT as the test was run with the same patchset id (115243-1) 2 days ago we got the artifacts of old 23/11 artifacts on LF backend, which looks good
the only unexplained element is why we have logs showing a SDC distribution and a reporting fully green.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>