[AAF] Externalizes init data out from aaf-cass
Instead of keep initialization data into the docker image, we move them
to the chart.
This will simplify adding / removing certificates as no image release
will be necessary, but only a change in OOM which can be directly
tested.

This reverts commit aa950d2ea4afa900c69eb622a776c8830a1c4e33.
Reason for revert: Changed the initial commit in order to make it really
work.

Issue-ID: OOM-2586
Change-Id: I1091d8940a8499503c72bd92ded9a54f08091ee3
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>