[SDNC] Tuning fixes for SDNC
Corrected issue in values.yaml that was causing OpenDaylight akka
configuration file override not to be loaded properly.  As a result,
we are seeing memory-related exceptions (circuit-breaker exception), whic
seems to be contributing to a memory leak error.

Also, updated configuration to enhance logging so that summary of
messages in/out (aka 'audit log') is written for debugging purposes.

Issue-ID: SDNC-1430
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: I2a3950ecde1c53f8e68b11bbc9c0ddc136f67b54