Update patch set 5
Patch Set 5: Code-Review+1

Experimental Azure shows as issue with the basic_vm test Heat template, and the OpenStack network.
This is not really an orchestration issue, rather a deficit in the environment.

Non-experimental Azure shows an issue with basic_cnf, which cannot be observed in experimental, and no logs are available.

It looks like SO code is working properly, and we have some environment issues.
I`d vote, that we merge this code for Guilin branch.

Patch-set: 5
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1