Update patch set 1
Patch Set 1: Code-Review-1

Hello,
thank you for the patch.

IMHO, fluentd is not part of ONAP but part of the "platform" where ONAP is installed.
So it shouldn't be in ONAP charts.
Also, fluentd is one tool amongs other for log processing, so I believe we shouldn't mandate choice.
As long as your logs are on STDOUT, it should be sufficient.

I'll take a look later this week on on multicloud part

Patch-set: 1
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1