[GENERAL] Fix setting of HELM_VER variable
It turns out that helm v2 and helm v3 uses different format for the
helm version command. If we use v3 version HELM_VER variable is empty
when run on a system with helm v2.

To fix this let's add a fallback and if we failed to obtain correct
version information using v3 format then fallback to v2 one.

Issue-ID: OOM-1
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I87757c9eadedfe197ec31f5c2814fa9eb894329a