Update patch set 7
Patch Set 7:

Thanks for this, Sylvain! 

Once this is merged, can we cherry pick this to guilin to address the expired cert there as well?  I checked the code and verified that cds-ui supports $KEYSTORE for location of keystore and $PASSPHRASE for location of credentials file there as well, so this same fix should work in Guilin.

Patch-set: 7
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>