Update patch set 5
Patch Set 5:

> Patch Set 5: Verified-1
> 
> Build Failed 
> 
> https://jenkins.onap.org/job/oom-guilin-verify-helm2/76/ : FAILURE
> 
> No problems were identified. If you know why this problem occurred, please add a suitable Cause for it. ( https://jenkins.onap.org/job/oom-guilin-verify-helm2/76/ )
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-guilin-verify-helm2/76
> 
> https://jenkins.onap.org/job/oom-frankfurt-verify-helm2/40/ : FAILURE
> 
> No problems were identified. If you know why this problem occurred, please add a suitable Cause for it. ( https://jenkins.onap.org/job/oom-frankfurt-verify-helm2/40/ )
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-frankfurt-verify-helm2/40
> 
> https://jenkins.onap.org/job/oom-master-image-verify/1204/ : SUCCESS
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-master-image-verify/1204
> 
> https://jenkins.onap.org/job/oom-master-commit-lint/655/ : SUCCESS
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-master-commit-lint/655
> 
> https://jenkins.onap.org/job/oom-master-verify-helm3.3/104/ : SUCCESS
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-master-verify-helm3.3/104
> 
> https://jenkins.onap.org/job/oom-master-verify-helm3.5/104/ : SUCCESS
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-master-verify-helm3.5/104

The Helm2 verification steps are failing because of a problem with CMPv2. few days ago, a change was made to a template.  The change introduced the use of the "mustFirst" function, which is not available in Helm 2.   It is no longer possible to build the OOM charts on the master branch using Helm 2.

Here is a log entry showing the problem:

02:30:10 ==> Linting cmpv2Certificate
02:30:10 [INFO] Chart.yaml: icon is recommended
02:30:10 [ERROR] templates/: parse error in "cmpv2Certificate/templates/_certServiceClient.tpl": template: cmpv2Certificate/templates/_certServiceClient.tpl:95: function "mustFirst" not defined
02:30:10 
02:30:10 Error: 1 chart(s) linted, 1 chart(s) failed

You can also see this verification failure on other submissions unrelated to DCAE.

If the intention is for Honolulu to build on Helm 3 only, then the helm2 verification jobs should be disabled for the master branch.   If the intention is for Honolulu to build on both Helm 2 and Helm 3, then it will be necessary to change the CMPv2 template so that it is compatible with Helm 2.

Patch-set: 5