[AAI][SPARKY] Disable Portal Spring Profile
In order for the aai ui to work properly in honolulu
the portal profile is being turned off. Portal
does not have the support needed and the future of
the component in ONAP is uncertain. The AAI UI will
still work in the portal with the profile turned off.
It will just not use its authentication integration
due to numerous issues dating back to Frankfurt.

Issue-ID: AAI-2905
Signed-off-by: wr148d <wr148d@att.com>
Change-Id: I7e120a16ced18b174c20bd9098883cd0e0e5d717