Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> we would need after to add it into ci-management right?

If the question is for me, I am not sure what to answer but I would say that the decision is up to you.
As far as I am concerned, I think it is already nice to have the possibility to run these profiles locally with a shared parameters configuration different from pylint/autopep8 defaults.
This is the configuration I use in ODL but feel free to propose anything different if it makes sense to you for ONAP.

Patch-set: 3
Reviewer: Gerrit User 3685 <3685@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>