Update patch set 2
Patch Set 2: Code-Review-1

FYI - the ORAN team discovered an issue with how we're installing their A1 adaptor, so I'm in the process of doing a new release build on the sdnc docker to pick up that change.  It should be ready today. I was planning to push a new patchset to this change with that updated version, as opposed to creating a separate review.

Patch-set: 2
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1