Update patch set 5
Patch Set 5:

Patchset 4 verified properly.  The gating results look good, but it turns out that the gating process did not enable the new dcaegen2-services charts.   In the onap values.yaml, we added a flag for dcaegen2-services.enabled, and by default it is set to false.  The override file for the gating process needs to be changed to set the flag to true.

In patchset 5, I have submitted changes to oom/kubernetes/onap/resources/environments/*.yaml and
oom/kubernetes/onap/resources/overrides/*.yaml.  I've added an entry for dcaegen2-services everywhere there is an entry for dcaegen2, and I've set the "enabled" flag for dcaegen2-services to match the
setting for the dcaegen2 "enabled" flag.

If gating is using the files in onap/resources, then these changes should cause the new dcaegen2-services charts to be deployed.

Patch-set: 5