Update patch set 5
Patch Set 5:

Patchset 5 uses the new charts to deploy hv-ves, prh, tcagen2, and ves as expected.  All four come up and become ready.  The separate "healthcheck" cases for hv-ves and ves succeed.  The pnfregistrate "smoketest" succeeds.  The 5gbulkpm fails, but I am not sure why.  (I'm not sure how dcae is involved
in the test.)
The dcaegen2 healthcheck fails.  This is because it is detecting a mongo instance being deployed dynamically by one of the other healthcheck/smoketests and it is present but not reporting as ready.  We observed this previously.

Patch-set: 5