Update patch set 5
Patch Set 5:

> Patch Set 4: Code-Review-1
> 
> APP_PORT environment variable is not set, using default.
> BASIC_AUTH environment variable is not set, using default.
> Traceback (most recent call last):
>   File "server.py", line 28, in 
>     import proto.CommandExecutor_pb2_grpc as CommandExecutor_pb2_grpc
>   File "/opt/app/onap/python/proto/CommandExecutor_pb2_grpc.py", line 4, in 
>     import CommandExecutor_pb2 as CommandExecutor__pb2
> ModuleNotFoundError: No module named 'CommandExecutor_pb2'

The unset variable statements are just warnings (not related to the ModuleNotFoundError).
The ModuleNotFoundError is due to a typo on the import.  In Java, the compiler would have
caught that, but no such luck with Python unfortunately.  We'll need to re-release CDS
dockers with the fix.  Should be ready tomorrow (I already pushed the fix - now it's just
a matter of getting it reviewed/merged, rebuilding and re-releasing dockers).

Patch-set: 5
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>