Update patch set 6
Patch Set 6:

> Patch Set 5: Code-Review-1
> 
> https://gitlab.com/Orange-OpenSource/lfn/onap/xtesting-onap/-/pipelines/264457456
> 
> 2 deployments failing
> onap-cds-command-executor 	0/1 	1 	0
> onap-cds-py-executor 	0/1 	1 	0
> helm test OK
> 
> all smoke tests failing at SO/SDNC
> oe"}},"instanceReferences":{"serviceInstanceId":"c35a01e2-c50c-49b1-b1c9-e0853da0c495","serviceInstanceName":"basic_vm-5f26fa6f-5851-4e2c-a218-54f6f5f1a662"},"requestStatus":{"requestState":"IN_PROGRESS","statusMessage":"STATUS: Error Source: SDNC, Error Message: Error from SDNC: No service logic active for generic-resource-api: 'service-topology-operation' RETRY STATUS: Retry 3/5 will be started in PT40S ROLLBACK STATUS: Error from SDNC: No service logic active for generic-resource-api: 'service-topology-operation'","percentProgress":100,"timestamp":"Wed, 03 Mar 2021 02:32:37 GMT"}}}

Would it be possible to get the logs from the dev-sdnc-0 pod?  I'm not seeing an issue loading the DGs when I start locally, so I would really need to see the log in gating to diagnose this issue.

On the issue with preload data not found, I'd also need to see the log.  SDNC doesn't have any preload data preloaded, so I suspect that either (a) there was an error of some sort in processing the preload or (b) if this is a multiple-node cluster (replicaCount > 1), the issue is probably due to an AKKA error on startup.  Our helm chart defaults to replicaCount 1, so (b) would only happen if you're overriding that

Patch-set: 6
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>