[COMMON] Add a checkbashisms tox profile
to identify potential bashisms in scripts
and ensure compatibility with dash and other POSIX shells
as often found in MAC OS X and BSD or UNIX systems.

Though bash is a good interactive user shell,
many UN*X systems and GNU+Linux distributions now prefers Almquist
or Kornshell variants for scripting in order to improve the system
performance and maintenability.

https://wiki.ubuntu.com/DashAsBinSh
https://mywiki.wooledge.org/Bashism
https://en.wikipedia.org/wiki/Almquist_shell

Issue-ID: OOM-2643
Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>
Change-Id: Ibfa114f5c4797e526edc6f59397117e093868c24