Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> by the way, you've got helm hooks that may break OOM cleanup :(
> for example: https://gitlab.com/Orange-OpenSource/lfn/onap/onap_oom_automatic_installation/-/jobs/1087191331
> 
> please be sure your hook ends well when we're on corner case (database not there anymore especially)

I can't see exactly what's causing the problem in your cleanup case, but it's clear from the code that if Consul is not running at the time the hook executes, the hook will keep executing, waiting for Consul to come up.  (I missed this because I always use helm undeploy, which uninstalls in reverse alphabetical order, so that dcaegen2-services is always uninstalled before consul.)  I've raised an issue (https://jira.onap.org/browse/DCAEGEN2-2669) for this.

For the short term, I believe it's reasonable to simply remove the hook.  The hook is simply removing some configuration information from Consul.  It's harmless (at least for the 4 microservices we are
currently deploying), since a new deployment of a microservice after it has been deleted would simply overwrite any existing configuration.

Patch-set: 9