[VID] Obfuscate truststore password at the entrypoint
VID unfortunately expects truststore password to be obfuscated instead
of getting it in plain text.
Current solution that we had in the master branch was only a temporary
fix which hardcoded obfuscated password in our charts.

This patch introduces a proper solution that uses jetty-util to
obfuscate the password before starting VID application.

Issue-ID: OOM-1
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I391c29b6552121fd0b574a1046b4c1551c1a927a