Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> the error on basic_onboard is a new exception when retrieving the list of CERTIFIED VSP
> this could be linked to the update and a change of behavior not taken into account in the sdk
>   File "/usr/lib/python3.8/site-packages/onapsdk/sdc/vsp.py", line 246, in load_status
>     and item_details['results'][-1]['status'] == const.CERTIFIED):
> IndexError: list index out of range
> 
> https://logs.onap.org/onap-integration/gating//117348-9/smoke-usecases/basic_onboard/basic_onboard/pythonsdk.debug.log.gz
> 
> or an init issue not detected so far because other tests were creating a VSP so the list was never empty - healthdist is Fail (no idea why the status is green on the dashboard, all the other robot status tests are correct, I do not see any risk as there were no several concurrent deployment)
> 
> basic_cnf is ok (using the same mechanism but basic_onboard is the first one, so the list of VSP is empty)
> basic_network and vm are failing at the last stage of instantiation due to openstack certificates, it has been fixed. Rebasing and retesting shall fix this issue

Now that logs are better on SDC, it's easier to read :=)

What happens :
Basic Onboard is "waiting" to see if it's component is there (Not sure why as the test has not yet even created the VSP but only created the VLM ??) to do so, it lists the VSP.
at the same time, another test is onboarding a vFW VSP, and bad luck, basic onboard is trying to read an VSP at the exact moment that it is being created, so all objects are not yet there. In short the get list of VSP is already returning the object ID, but the object itself is still under creation so when basic onboard tries to do a get item, it gets an empty response.

this is visible in onboarding backend logs (first thread is still creating the object, when another one come and tries to get it)

2021-03-19T18:37:11.002Z	[qtp580718781-14]	INFO 	c.a.z.c.l.i.ZusammenSLF4JLoggerFactoryImpl	RequestID=d8892796-6791-41ee-a163-846f5812275c	InstanceID=296f8810-7dd0-48cd-a72d-0b79d5f05c6f	PartnerName=cs0008	ServerFQDN=onap-sdc-onboarding-be-778898d489-87qbc	ServerIpAddress=10.233.74.252	ServiceName=POST: /onboarding-api/v1.0/vendor-software-products	create item version - item:61e07163ead54f7d98fa0bc0bb1774de version:51bd81f8f24c45b1add0a5598139697b

2021-03-19T18:37:11.063Z	[qtp580718781-11]	INFO 	o.o.s.i.rest.services.VersionsImpl	ResponseCode=200	RequestID=4bce6137-4845-4224-b246-8123613c805a	InstanceID=296f8810-7dd0-48cd-a72d-0b79d5f05c6f	ElapsedTime=110	EndTimestamp=2021-03-19T18:37:11.063Z	PartnerName=cs0008	ServerFQDN=onap-sdc-onboarding-be-778898d489-87qbc	ClientIPAddress=10.233.68.42	StatusCode=COMPLETE	ServerIpAddress=10.233.74.252	ServiceName=GET: /onboarding-api/v1.0/items/61e07163ead54f7d98fa0bc0bb1774de/versions	EntryTimestamp=2021-03-19T18:37:10.953Z	InvocationID=30ab460d-91da-4ac7-b74f-b98417f6510a	ResponseDesc=OK	

again a retry in the test should fix this, also I don't get the 'logic' of the test, as it tries to identify if a VSP it needs is already onboarded, and does a list then get all items from the list, but list output itself is enough to determine if its there or not ? see output from the list :

the test is searching for :
2021-03-19 18:36:45,823 DEBUG 159:__init__.py(1) - check if Vsp basic_onboard_DdLWxI_VSP exists in SDC

the output of get VSPs from SDC is : 
[SDC][get Vsps] response: {"results":[{"name":"vFW f76f253d-c67d","description":"vendor software product","vendorName":"132ffcb6-13a7-4481-85e2","vendorId":"0cb2806d207b4af8afbacfed0d179162","onboardingMethod":"NetworkPackage","id":"61e07163ead54f7d98fa0bc0bb1774de","owner":"cs0008","status":"ACTIVE"},{"name":"vFW 0fe71258-a1fd","description":"vendor software product","vendorName":"8da71b96-d300-4624-8d39","vendorId":"4840fbad90c34b63893c37beabb475ec","onboardingMethod":"NetworkPackage","id":"2cd23ad1a381481e8e1f63af8702e0e6","owner":"cs0008","status":"ACTIVE"}],"listCount":2}

I can understand that another thread in the test may be creating the needed VSP but still
I don't see the point of fetching list items that do not match the name the test is searching for ? or am i missing something ?

Patch-set: 9