Update patch set 12
Patch Set 12:

> Patch Set 12: Code-Review-1
> 
> all the E2E tests involving the full chain SDC - SO - SDNC -AAI seem to be fail due to tiemout on SO requests (A la carte BPMN)
> there is no error reported by SO (each request in is IN  PROGRESS during 15 minutes)
> however if we look into the pod logs from the cluster snapshot executed after the tests (k8s-teardown), we can see exception in SDNC
> 
> see https://logs.onap.org/onap-integration/gating//118284-12/infrastructure-healthcheck/k8s-teardown/kubernetes-status/pod-onap-sdnc-0-sdnc-logs.html

The exception is unrelated to SO:
org.onap.ccsdk.sli.core.sli.SvcLogicException: Exception while posting http request to client URI is not absolute
	at org.onap.ccsdk.sli.plugins.restapicall.RestapiCallNode.sendHttpRequest(RestapiCallNode.java:855) ~[bundleFile:?]
	at org.onap.ccsdk.sli.plugins.restapicall.RestapiCallNode.sendRequest(RestapiCallNode.java:501) [bundleFile:?]
	at org.onap.ccsdk.sli.plugins.restapicall.RestapiCallNode.sendRequest(RestapiCallNode.java:474) [bundleFile:?]
	at jdk.internal.reflect.GeneratedMethodAccessor105.invoke(Unknown Source) ~[?:?]
	at jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(Unknown Source) ~[?:?]

This is actually an issue in A1 policy management (see https://jira.onap.org/browse/CCSDK-3227).  By default, it has no non-RT RICs configured, and so should not be sending a message to SDNC.  However, it is sending an invalid message, which in turn causes this error.  This shouldn't hurt anything, but we will address it so that it doesn't flood the logs and obscure true errors.

Patch-set: 12
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>