Update patch set 12
Patch Set 12:

> Patch Set 12: Code-Review-1
> 
> network name gen fails on the last tests

I see the problem.  We changed to use the integration image, and forgot that the integration image defines an ENTRYPOINT that assumes the code to be executed is in /app.jar - which is not the case here.  Simple fix - I'll push an update

Patch-set: 12
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>